Add parse_binary_modules to WastParserOptions.
Enabled by default, this option implements the current behavior of WABT
when it encounters "(module binary ...)" script modules. That is, it
parses the binary module, reporting any errors that it sees.

When disabled, the binary module will be passed back to the WAST parser
as unvalidated raw bytes. This allows for the use of WABT in places
which want to provide invalid modules to module loaders (e.g. spec test
runners).

Fixes https://github.com/WebAssembly/wabt/issues/2561.