Fix for empty CMAKE_OSX_SYSROOT (#2634)
This seems to have started happening in on the macos-latest CI machine.
Seems better to check for empty string since apparently that is what it
gets set to in cmake 4.0:

https://cmake.org/cmake/help/latest/variable/CMAKE_OSX_SYSROOT.html:

"Changed in version 4.0: The default is now empty. Previously a default
was computed based on the CMAKE_OSX_DEPLOYMENT_TARGET or the host
platform."

It looks like there is another issue that is causing some tests to fail
on macos-latest (`sigaltstack: Invalid argument`) but I proposal that we
fix that in a separate PR.