Add support for (ref ...) / (ref null ...) constructs (#2562)
This large patch adds function references support for those parts of the
wabt library, which already present in the code. The current code is
designed for an older (outdated) variant of the function references
proposal.

Key changes:
- Var variables can optionally store type data. This way there is no
need for a Var/Type field pair in many cases.
- CallRef now needs to explicitly provide its reference type (no auto
detection)
- Tracking local reference initialization (non nullable refs must be set
before using them)
- Supporting type equality comparisons in the validator and in the
interpreter
- Adds a code for reading/writing 33 bit integers (leb format)
- Remove EndTypeSection in shared validator, types can be validated
earlier
- Improve named reference resolving in the text parser

Co-authored-by: Zoltan Herczeg <hzmester@freemail.hu>