wast-parser: add support for '(module definition ...)'
WASM v3 spec tests added the '(module definition ...)' construct to
WAST. This tells the test runner that the module should not be
instantiated immediately. The '(module instance ...)' construct (not yet
supported in wabt) instructs the test runner to instantiate the module.

This change adds an `is_definition` field to ScriptModule. The field
reflects whether the `definition` keyword was encountered after the
`module` keyword.

When WastParser encounters a TextModule in a script command, it now
returns a ScriptModule with the corresponding ScriptTextModule if the
module is definition-only.

This change allows wabt to parse WASM v3 spec tests that use the
`(module definition ...)` construct (specifically, `memory.wast`), so
that test runners built on wabt can support WASM v3 spec tests.